# backend/src/swagger/swagger.json
{
	"definitions": {
		"genericObject": {
			"description": "",
			"properties": {},
			"type": "object"
		},
		"BadRequestError": {
			"description": "Represents a BAD REQUEST error. The request could not be understood by the\nserver due to malformed syntax. The client SHOULD NOT repeat the request\nwithout modifications.",
			"properties": {
				"message": {
					"type": "string",
					"description": ""
				},
				"statusCode": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"stack": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"message",
				"statusCode",
				"name"
			]
		}
	},
	"info": {
		"license": {
			"name": "ISC"
		},
		"title": "backend",
		"version": "1.0.0"
	},
	"paths": {
		"/testGET": {
			"get": {
				"operationId": "ServiceTestGET",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/genericObject"
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"$ref": "#/definitions/BadRequestError"
						}
					}
				},
				"description": "",
				"parameters": [
					{
						"description": "",
						"in": "query",
						"name": "key",
						"required": true,
						"type": "string"
					},
					{
						"description": "",
						"in": "query",
						"name": "value",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			}
		},
		"/testPOST": {
			"post": {
				"operationId": "ServiceTestPOST",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"$ref": "#/definitions/BadRequestError"
						}
					}
				},
				"description": "",
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "query",
						"name": "test",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			}
		},
		"/testGetDb": {
			"get": {
				"operationId": "ServiceTestDBGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"$ref": "#/definitions/BadRequestError"
						}
					}
				},
				"description": "",
				"parameters": [
					{
						"description": ":string - username entered into login page",
						"in": "query",
						"name": "username",
						"required": true,
						"type": "string"
					},
					{
						"description": ":string - password entered into login page (clear text no encryption)",
						"in": "query",
						"name": "password",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/postTableDB": {
			"post": {
				"operationId": "ServicePostTableDB",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"$ref": "#/definitions/BadRequestError"
						}
					}
				},
				"description": "",
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": ": any - 2d array which contains data of every cell in the spreadsheet",
						"in": "body",
						"name": "dataArray",
						"required": true,
						"schema": {
							"type": "object"
						}
					}
				]
			}
		}
	},
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "access_token",
			"in": "query"
		}
	},
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	]
}